requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
matplotlib
seaborn
python-binance
ta
tqdm
requests
websocket-client
python-telegram-bot==20.3
python-dotenv
pytelegrambotapi
